requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
qdrant-client==1.8.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
pydantic-settings==2.1.0
black==23.12.1
isort==5.13.2
flake8==6.0.0
mypy==1.8.0
pre-commit==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.4.0
limits==3.10.0